Function mentat_ffi::in_progress_builder_retract_string

#[no_mangle]

pub unsafe extern "C" fn in_progress_builder_retract_string<'a, 'c>(
    builder: *mut InProgressBuilder<'a, 'c>, 
    entid: c_longlong, 
    kw: *const c_char, 
    value: *const c_char
) -> *mut ExternResult

Uses builder to retract value for kw on entity entid.

Errors

If entid is not present in the store. If kw is not a valid attribute in the store. If the :db/type of the attribute described by kw is not :db.type/string.