Struct mentat_ffi::EntityBuilder

pub struct EntityBuilder<T> where
    T: BuildTerms,  { /* fields omitted */ }

Methods

impl<T> EntityBuilder<T> where
T: BuildTerms,

pub fn finish(self) -> (T, Either<KnownEntid, Rc<TempId>>)

pub fn add<V>(&mut self, a: KnownEntid, v: V) -> Result<(), Error> where
V: IntoThing<Either<TypedValue, Rc<TempId>>>,

pub fn retract<V>(&mut self, a: KnownEntid, v: V) -> Result<(), Error> where
V: IntoThing<Either<TypedValue, Rc<TempId>>>,

impl<'a, 'c> EntityBuilder<InProgressBuilder<'a, 'c>>

pub fn add_kw<V>(&mut self, a: &Keyword, v: V) -> Result<(), Error> where
V: IntoThing<Either<TypedValue, Rc<TempId>>>,

pub fn retract_kw<V>(&mut self, a: &Keyword, v: V) -> Result<(), Error> where
V: IntoThing<Either<TypedValue, Rc<TempId>>>,

pub fn transact(self) -> (InProgress<'a, 'c>, Result<TxReport, Error>)

Build the terms from this builder and transact them against the current InProgress. This method always returns the InProgress -- failure doesn't imply an automatic rollback.

pub fn commit(self) -> Result<TxReport, Error>

Transact the contents of the builder and commit the InProgress. If any step fails, roll back. Return the TxReport.