Struct mentat_ffi::InProgress

pub struct InProgress<'a, 'c> { /* fields omitted */ }

Represents an in-progress, not yet committed, set of changes to the store. Call commit to commit your changes, or rollback to discard them. A transaction is held open until you do so. Your changes will be implicitly dropped along with this struct.

Methods

impl<'a, 'c> InProgress<'a, 'c>

pub fn builder(self) -> InProgressBuilder<'a, 'c>

pub fn use_caching(&mut self, yesno: bool)

Choose whether to use in-memory caches for running queries.

pub fn transact_builder(
&mut self,
builder: TermBuilder
) -> Result<TxReport, Error>

If you only have a reference to an InProgress, you can't use the easy builder. This exists so you can make your own.

pub fn transact_terms<I>(
&mut self,
terms: I,
tempid_set: InternSet<TempId>
) -> Result<TxReport, Error> where
I: IntoIterator<Item = Term<Either<KnownEntid, Rc<TempId>>, Either<TypedValue, Rc<TempId>>>>,

pub fn transact_entities<I, V>(
&mut self,
entities: I
) -> Result<TxReport, Error> where
I: IntoIterator<Item = Entity<V>>,
V: TransactableValue,

pub fn transact<B>(&mut self, transaction: B) -> Result<TxReport, Error> where
B: Borrow<str>,

pub fn import<P>(&mut self, path: P) -> Result<TxReport, Error> where
P: AsRef<Path>,

pub fn rollback(self) -> Result<(), Error>

pub fn commit(self) -> Result<(), Error>

pub fn cache(
&mut self,
attribute: &Keyword,
cache_direction: CacheDirection,
cache_action: CacheAction
) -> Result<(), Error>