Struct mentat_ffi::QueryInputs

pub struct QueryInputs { /* fields omitted */ }

Define the inputs to a query. This is in two parts: a set of values known now, and a set of types known now. The separate map of types is to allow queries to be algebrized without full knowledge of the bindings that will be used at execution time. When built correctly, types is guaranteed to contain the types of values -- use QueryInputs::new or QueryInputs::with_values to construct an instance.

Methods

impl QueryInputs

pub fn with_value_sequence(vals: Vec<(Variable, TypedValue)>) -> QueryInputs

pub fn with_type_sequence(types: Vec<(Variable, ValueType)>) -> QueryInputs

pub fn with_values(values: BTreeMap<Variable, TypedValue>) -> QueryInputs

pub fn new(
types: BTreeMap<Variable, ValueType>,
values: BTreeMap<Variable, TypedValue>
) -> Result<QueryInputs, Error>